refactor(surpriseexam): Modularize Surprise Exam random event - Moved OSRSItemRelationshipSystem to surpriseexam module - Moved PatternRelation to surpriseexam module - Added Surprise Exam config option - Refactored Surprise Exam code into its own module 	- Added SurpriseExamHelper and SurpriseExamOverlay
