+++ src/main/java/randomeventsolver/RandomEventSolverPlugin.java
@@ -96,70 +96,7 @@ public class RandomEventSolverPlugin extends Plugin
 	private RandomEventSolverItemOverlay itemOverlay;
 
 	@Getter
-	private ImmutableSet<RandomEventItem> patternCardAnswers;
-
-	@Getter
-	private ImmutableSet<Widget> patternCardAnswerWidgets;
-
-	@Getter
-	private RandomEventItem patternNextAnswer;
-
-	@Getter
-	private Widget patternNextAnswerWidget;
-
-	private OSRSItemRelationshipSystem relationshipSystem;
-
-	@Getter
 	private ImmutableList<Widget> beehiveAnswerWidgets;
-
-	private static final int[] PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_MODELS = {
-		InterfaceID.PatternCards.CARD_0,
-		InterfaceID.PatternCards.CARD_1,
-		InterfaceID.PatternCards.CARD_2,
-		InterfaceID.PatternCards.CARD_3,
-		InterfaceID.PatternCards.CARD_4,
-		InterfaceID.PatternCards.CARD_5,
-		InterfaceID.PatternCards.CARD_6,
-		InterfaceID.PatternCards.CARD_7,
-		InterfaceID.PatternCards.CARD_8,
-		InterfaceID.PatternCards.CARD_9,
-		InterfaceID.PatternCards.CARD_10,
-		InterfaceID.PatternCards.CARD_11,
-		InterfaceID.PatternCards.CARD_12,
-		InterfaceID.PatternCards.CARD_13,
-		InterfaceID.PatternCards.CARD_14
-	};
-
-	private static final int[] PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_SELECTS = {
-		InterfaceID.PatternCards.SELECT_0,
-		InterfaceID.PatternCards.SELECT_1,
-		InterfaceID.PatternCards.SELECT_2,
-		InterfaceID.PatternCards.SELECT_3,
-		InterfaceID.PatternCards.SELECT_4,
-		InterfaceID.PatternCards.SELECT_5,
-		InterfaceID.PatternCards.SELECT_6,
-		InterfaceID.PatternCards.SELECT_7,
-		InterfaceID.PatternCards.SELECT_8,
-		InterfaceID.PatternCards.SELECT_9,
-		InterfaceID.PatternCards.SELECT_10,
-		InterfaceID.PatternCards.SELECT_11,
-		InterfaceID.PatternCards.SELECT_12,
-		InterfaceID.PatternCards.SELECT_13,
-		InterfaceID.PatternCards.SELECT_14
-	};
-
-	private static final int[] PATTERNNEXT_INTERFACEIDS_INITIAL_PATTERN = {
-		InterfaceID.PatternNext._0,
-		InterfaceID.PatternNext._1,
-		InterfaceID.PatternNext._2
-	};
-
-	private static final int[] PATTERNNEXT_INTERFACEIDS_CHOICES = {
-		InterfaceID.PatternNext.SELECT_0,
-		InterfaceID.PatternNext.SELECT_1,
-		InterfaceID.PatternNext.SELECT_2,
-		InterfaceID.PatternNext.SELECT_3
-	};
 
 	private final String FREAKY_FORESTER_REGEX = "Could you kill a pheasant with (?<numberOfTails>\\d+) tails";
 	private final Pattern FREAKY_FORESTER_PATTERN = Pattern.compile(FREAKY_FORESTER_REGEX, Pattern.CASE_INSENSITIVE);
@@ -187,11 +124,6 @@ public class RandomEventSolverPlugin extends Plugin
 	{
 		this.overlayManager.add(overlay);
 		this.overlayManager.add(itemOverlay);
-		this.patternCardAnswers = null;
-		this.patternCardAnswerWidgets = null;
-		this.patternNextAnswer = null;
-		this.patternNextAnswerWidget = null;
-		this.relationshipSystem = new OSRSItemRelationshipSystem();
 	}
 
 	@Override
@@ -199,11 +131,6 @@ public class RandomEventSolverPlugin extends Plugin
 	{
 		this.overlayManager.remove(overlay);
 		this.overlayManager.remove(itemOverlay);
-		this.patternCardAnswers = null;
-		this.patternCardAnswerWidgets = null;
-		this.patternNextAnswer = null;
-		this.patternNextAnswerWidget = null;
-		this.relationshipSystem = null;
 		this.exerciseMatsAnswerList.clear();
 		this.exerciseMatsMultimap.clear();
 		this.exerciseVarbitMatMultimap.clear();
@@ -284,86 +211,6 @@ public class RandomEventSolverPlugin extends Plugin
 	public void onWidgetLoaded(WidgetLoaded widgetLoaded)
 	{
 		// log.debug("Widget loaded with group ID: {}", widgetLoaded.getGroupId());
-		if (widgetLoaded.getGroupId() == InterfaceID.PATTERN_CARDS)
-		{
-
-			this.clientThread.invokeLater(() -> {
-				Widget examHintWidget = this.client.getWidget(InterfaceID.PatternCards.HINT);
-				if (examHintWidget != null)
-				{
-					String examHint = examHintWidget.getText();
-					log.debug("Exam hint widget loaded with text: {}", examHint);
-					if (examHint != null && !examHint.isEmpty())
-					{
-						List<RandomEventItem> answerItems = this.relationshipSystem.findItemsByHint(examHint, this.getPatternCardMap().values().asList(), 3);
-						log.debug("Found answer items for exam hint '{}': {}", examHint, answerItems);
-						if (answerItems.size() >= 3)
-						{
-							this.patternCardAnswers = ImmutableSet.copyOf(answerItems);
-							this.patternCardAnswerWidgets = answerItems.subList(0, 3).stream()
-								.map(item -> {
-									Integer interfaceID = this.getKeyForValue(this.getPatternCardMap(), item);
-									Widget interfaceWidget = this.client.getWidget(interfaceID);
-									return interfaceWidget != null ? this.getPatternCardSelectionWidgetFromModel(interfaceWidget) : null;
-								})
-								.filter(Objects::nonNull)
-								.collect(ImmutableSet.toImmutableSet());
-							log.debug("Pattern card answers set to: {}", this.patternCardAnswers);
-							log.debug("Pattern card answer widgets set to: {}", this.patternCardAnswerWidgets);
-						}
-						else
-						{
-							log.warn("Found {} items for exam hint '{}', expected 3.", answerItems.size(), examHint);
-							this.patternCardAnswers = null;
-							this.patternCardAnswerWidgets = null;
-						}
-					}
-					else
-					{
-						log.warn("Exam hint widget text is empty or null.");
-						this.patternCardAnswers = null;
-					}
-				}
-			});
-		}
-
-		if (widgetLoaded.getGroupId() == InterfaceID.PATTERN_NEXT)
-		{
-			this.clientThread.invokeLater(() -> {
-				Widget whatsNextTextWidget = this.client.getWidget(InterfaceID.PatternNext.UNIVERSE_TEXT12);
-				if (whatsNextTextWidget != null)
-				{
-					String whatsNextText = whatsNextTextWidget.getText();
-					log.debug("What's next widget text loaded: {}", whatsNextText);
-					if (whatsNextText != null && !whatsNextText.isEmpty())
-					{
-						List<RandomEventItem> initialSelectionItems = this.getPatternNextInitialSelectionMap().values().asList();
-						List<RandomEventItem> choicesItems = this.getPatternNextChoicesMap().values().asList();
-						RandomEventItem answerItem = this.relationshipSystem.findMissingItem(initialSelectionItems, choicesItems);
-						if (answerItem != null)
-						{
-							this.patternNextAnswer = answerItem;
-							Integer interfaceID = getKeyForValue(this.getPatternNextChoicesMap(), answerItem);
-							this.patternNextAnswerWidget = interfaceID != null ? this.client.getWidget(interfaceID) : null;
-							log.debug("Pattern next answer set to: {}", this.patternNextAnswer);
-							log.debug("Pattern next answer widget set to: {}", this.patternNextAnswerWidget);
-						}
-						else
-						{
-							log.warn("No valid answer found for what's next text '{}'.", whatsNextText);
-							this.patternNextAnswer = null;
-							this.patternNextAnswerWidget = null;
-						}
-					}
-					else
-					{
-						log.warn("Next hint widget text is empty or null.");
-						this.patternNextAnswer = null;
-						this.patternNextAnswerWidget = null;
-					}
-				}
-			});
-		}
 
 		if (widgetLoaded.getGroupId() == InterfaceID.BEEHIVE)
 		{
@@ -486,20 +333,6 @@ public class RandomEventSolverPlugin extends Plugin
 	public void onWidgetClosed(WidgetClosed widgetClosed)
 	{
 		// log.debug("Widget closed with group ID: {}", widgetClosed.getGroupId());
-		if (widgetClosed.getGroupId() == InterfaceID.PATTERN_CARDS)
-		{
-			log.debug("Pattern cards widget closed, resetting pattern card answers.");
-			this.patternCardAnswers = null;
-			this.patternCardAnswerWidgets = null;
-		}
-
-		if (widgetClosed.getGroupId() == InterfaceID.PATTERN_NEXT)
-		{
-			log.debug("Pattern next widget closed, resetting pattern next answer.");
-			this.patternNextAnswer = null;
-			this.patternNextAnswerWidget = null;
-		}
-
 		if (widgetClosed.getGroupId() == InterfaceID.BEEHIVE)
 		{
 			log.debug("Beehive widget closed, resetting beehive answer widgets.");
@@ -884,113 +717,10 @@ public class RandomEventSolverPlugin extends Plugin
 		return configManager.getConfig(RandomEventSolverConfig.class);
 	}
 
-	private ImmutableMultimap<Integer, RandomEventItem> getPatternCardMap()
-	{
-		ImmutableMultimap.Builder<Integer, RandomEventItem> builder = ImmutableMultimap.builder();
-		for (int patternCardSelectionInterfaceID : PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_MODELS)
-		{
-			int modelID = Objects.requireNonNull(this.client.getWidget(patternCardSelectionInterfaceID)).getModelId();
-			RandomEventItem randomEventItem = RandomEventItem.fromModelID(modelID);
-			if (randomEventItem == null)
-			{
-				log.warn("No RandomEventItem found for model ID: {}", modelID);
-				continue;
-			}
-			builder.put(patternCardSelectionInterfaceID, randomEventItem);
-		}
-		return builder.build();
-	}
-
-	private ImmutableMap<Integer, RandomEventItem> getPatternNextInitialSelectionMap()
-	{
-		ImmutableMap.Builder<Integer, RandomEventItem> builder = ImmutableMap.builder();
-		for (int patternNextInitialPatternInterfaceID : PATTERNNEXT_INTERFACEIDS_INITIAL_PATTERN)
-		{
-			int modelID = Objects.requireNonNull(this.client.getWidget(patternNextInitialPatternInterfaceID)).getModelId();
-			RandomEventItem randomEventItem = RandomEventItem.fromModelID(modelID);
-			if (randomEventItem == null)
-			{
-				log.warn("No RandomEventItem found for model ID: {}", modelID);
-				continue;
-			}
-			builder.put(patternNextInitialPatternInterfaceID, randomEventItem);
-		}
-		return builder.build();
-	}
-
-	private ImmutableMultimap<Integer, RandomEventItem> getPatternNextChoicesMap()
-	{
-		ImmutableMultimap.Builder<Integer, RandomEventItem> builder = ImmutableMultimap.builder();
-		for (int patternNextChoiceInterfaceID : PATTERNNEXT_INTERFACEIDS_CHOICES)
-		{
-			int modelID = Objects.requireNonNull(this.client.getWidget(patternNextChoiceInterfaceID)).getModelId();
-			RandomEventItem randomEventItem = RandomEventItem.fromModelID(modelID);
-			if (randomEventItem == null)
-			{
-				log.warn("No RandomEventItem found for model ID: {}", modelID);
-				continue;
-			}
-			builder.put(patternNextChoiceInterfaceID, randomEventItem);
-		}
-		return builder.build();
-	}
-
-	private Integer getKeyForValue(Multimap<Integer, RandomEventItem> map, RandomEventItem value)
-	{
-		for (Map.Entry<Integer, RandomEventItem> entry : map.entries())
-		{
-			if (entry.getValue().equals(value))
-			{
-				return entry.getKey();
-			}
-		}
-		return null;
-	}
-
-	public Widget getPatternCardSelectionWidgetFromModel(Widget modelWidget)
-	{
-		// This method retrieves the widget for the pattern card selection based on the model widget.
-		// PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_MODELS[i] = PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_SELECTS[i]
-		// So if modelWidget corresponds to PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_MODELS[i] then it should return PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_SELECTS[i]
-		for (int i = 0; i < PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_MODELS.length; i++)
-		{
-			if (modelWidget.getId() == PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_MODELS[i])
-			{
-				return this.client.getWidget(PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_SELECTS[i]);
-			}
-		}
-		log.warn("No matching selection widget found for model widget ID: {}", modelWidget.getId());
-		return null;
-	}
-
 	// Accounts for local instances too such as inside the pinball random event
 	private int getRegionIDFromCurrentLocalPointInstanced()
 	{
 		return WorldPoint.fromLocalInstance(this.client, this.client.getLocalPlayer().getLocalLocation()).getRegionID();
-	}
-
-	@Getter
-	@AllArgsConstructor
-	enum PatternCardAnswers
-	{
-
-		//		HINT_1("I feel like a fish out of water!", ImmutableSet.of(RandomEventItem.SALMON, RandomEventItem.TROUT, RandomEventItem.HARPOON)),
-//		HINT_2("Sea food, catch the food, spot the pattern.", ImmutableSet.of(RandomEventItem.SALMON, RandomEventItem.TROUT, RandomEventItem.HARPOON)),
-//		HINT_3("I'm fishing for answers.", ImmutableSet.of(RandomEventItem.SALMON, RandomEventItem.TROUT, RandomEventItem.HARPOON)),
-		HINT_4("The pen may be mightier than the sword, but against a dragon? I'll take a melee weapon.", ImmutableSet.of());
-
-		private final String hint;
-		private final ImmutableSet<RandomEventItem> answerModels;
-	}
-
-	@Getter
-	@AllArgsConstructor
-	enum PatternNextRelations
-	{
-		FRUITS(ImmutableSet.of(RandomEventItem.BANANA, RandomEventItem.ONION, RandomEventItem.PINEAPPLE, RandomEventItem.WATERMELON_SLICE)),
-		MAGIC(ImmutableSet.of(RandomEventItem.AIR_RUNE, RandomEventItem.EARTH_RUNE, RandomEventItem.FIRE_RUNE, RandomEventItem.STAFF));
-
-		private final ImmutableSet<RandomEventItem> models;
 	}
 
 	@Getter
