+++ src/main/java/randomeventsolver/randomevents/surpriseexam/SurpriseExamHelper.java
@@ -1,0 +1,312 @@
+package randomeventsolver.randomevents.surpriseexam;
+
+import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.ImmutableMultimap;
+import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Multimap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import javax.inject.Inject;
+import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
+import net.runelite.api.Client;
+import net.runelite.api.events.WidgetClosed;
+import net.runelite.api.events.WidgetLoaded;
+import net.runelite.api.gameval.InterfaceID;
+import net.runelite.api.widgets.Widget;
+import net.runelite.client.callback.ClientThread;
+import net.runelite.client.eventbus.Subscribe;
+import net.runelite.client.plugins.Plugin;
+import net.runelite.client.ui.overlay.OverlayManager;
+import randomeventsolver.data.RandomEventItem;
+
+@Slf4j
+public class SurpriseExamHelper extends Plugin
+{
+	@Inject
+	private Client client;
+
+	@Inject
+	private ClientThread clientThread;
+
+	@Inject
+	private OverlayManager overlayManager;
+
+	@Inject
+	private SurpriseExamOverlay overlay;
+
+	@Getter
+	private ImmutableSet<RandomEventItem> patternCardAnswers;
+
+	@Getter
+	private ImmutableSet<Widget> patternCardAnswerWidgets;
+
+	@Getter
+	private RandomEventItem patternNextAnswer;
+
+	@Getter
+	private Widget patternNextAnswerWidget;
+
+	private OSRSItemRelationshipSystem relationshipSystem;
+
+	private static final int[] PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_MODELS = {
+		InterfaceID.PatternCards.CARD_0,
+		InterfaceID.PatternCards.CARD_1,
+		InterfaceID.PatternCards.CARD_2,
+		InterfaceID.PatternCards.CARD_3,
+		InterfaceID.PatternCards.CARD_4,
+		InterfaceID.PatternCards.CARD_5,
+		InterfaceID.PatternCards.CARD_6,
+		InterfaceID.PatternCards.CARD_7,
+		InterfaceID.PatternCards.CARD_8,
+		InterfaceID.PatternCards.CARD_9,
+		InterfaceID.PatternCards.CARD_10,
+		InterfaceID.PatternCards.CARD_11,
+		InterfaceID.PatternCards.CARD_12,
+		InterfaceID.PatternCards.CARD_13,
+		InterfaceID.PatternCards.CARD_14
+	};
+
+	private static final int[] PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_SELECTS = {
+		InterfaceID.PatternCards.SELECT_0,
+		InterfaceID.PatternCards.SELECT_1,
+		InterfaceID.PatternCards.SELECT_2,
+		InterfaceID.PatternCards.SELECT_3,
+		InterfaceID.PatternCards.SELECT_4,
+		InterfaceID.PatternCards.SELECT_5,
+		InterfaceID.PatternCards.SELECT_6,
+		InterfaceID.PatternCards.SELECT_7,
+		InterfaceID.PatternCards.SELECT_8,
+		InterfaceID.PatternCards.SELECT_9,
+		InterfaceID.PatternCards.SELECT_10,
+		InterfaceID.PatternCards.SELECT_11,
+		InterfaceID.PatternCards.SELECT_12,
+		InterfaceID.PatternCards.SELECT_13,
+		InterfaceID.PatternCards.SELECT_14
+	};
+
+	private static final int[] PATTERNNEXT_INTERFACEIDS_INITIAL_PATTERN = {
+		InterfaceID.PatternNext._0,
+		InterfaceID.PatternNext._1,
+		InterfaceID.PatternNext._2
+	};
+
+	private static final int[] PATTERNNEXT_INTERFACEIDS_CHOICES = {
+		InterfaceID.PatternNext.SELECT_0,
+		InterfaceID.PatternNext.SELECT_1,
+		InterfaceID.PatternNext.SELECT_2,
+		InterfaceID.PatternNext.SELECT_3
+	};
+
+	@Override
+	protected void startUp() throws Exception
+	{
+		this.overlayManager.add(overlay);
+		this.patternCardAnswers = null;
+		this.patternCardAnswerWidgets = null;
+		this.patternNextAnswer = null;
+		this.patternNextAnswerWidget = null;
+		this.relationshipSystem = new OSRSItemRelationshipSystem();
+	}
+
+	@Override
+	protected void shutDown() throws Exception
+	{
+		this.overlayManager.remove(overlay);
+		this.patternCardAnswers = null;
+		this.patternCardAnswerWidgets = null;
+		this.patternNextAnswer = null;
+		this.patternNextAnswerWidget = null;
+		this.relationshipSystem = null;
+	}
+
+	@Subscribe
+	public void onWidgetLoaded(WidgetLoaded widgetLoaded)
+	{
+		// log.debug("Widget loaded with group ID: {}", widgetLoaded.getGroupId());
+		if (widgetLoaded.getGroupId() == InterfaceID.PATTERN_CARDS)
+		{
+
+			this.clientThread.invokeLater(() -> {
+				Widget examHintWidget = this.client.getWidget(InterfaceID.PatternCards.HINT);
+				if (examHintWidget != null)
+				{
+					String examHint = examHintWidget.getText();
+					log.debug("Exam hint widget loaded with text: {}", examHint);
+					if (examHint != null && !examHint.isEmpty())
+					{
+						List<RandomEventItem> answerItems = this.relationshipSystem.findItemsByHint(examHint, this.getPatternCardMap().values().asList(), 3);
+						log.debug("Found answer items for exam hint '{}': {}", examHint, answerItems);
+						if (answerItems.size() >= 3)
+						{
+							this.patternCardAnswers = ImmutableSet.copyOf(answerItems);
+							this.patternCardAnswerWidgets = answerItems.subList(0, 3).stream()
+								.map(item -> {
+									Integer interfaceID = this.getKeyForValue(this.getPatternCardMap(), item);
+									if (interfaceID == null)
+									{
+										log.warn("No interface ID found for item: {}", item);
+										return null;
+									}
+									Widget interfaceWidget = this.client.getWidget(interfaceID);
+									return interfaceWidget != null ? this.getPatternCardSelectionWidgetFromModel(interfaceWidget) : null;
+								})
+								.filter(Objects::nonNull)
+								.collect(ImmutableSet.toImmutableSet());
+							log.debug("Pattern card answers set to: {}", this.patternCardAnswers);
+							log.debug("Pattern card answer widgets set to: {}", this.patternCardAnswerWidgets);
+						}
+						else
+						{
+							log.warn("Found {} items for exam hint '{}', expected 3.", answerItems.size(), examHint);
+							this.patternCardAnswers = null;
+							this.patternCardAnswerWidgets = null;
+						}
+					}
+					else
+					{
+						log.warn("Exam hint widget text is empty or null.");
+						this.patternCardAnswers = null;
+					}
+				}
+			});
+		}
+
+		if (widgetLoaded.getGroupId() == InterfaceID.PATTERN_NEXT)
+		{
+			this.clientThread.invokeLater(() -> {
+				Widget whatsNextTextWidget = this.client.getWidget(InterfaceID.PatternNext.UNIVERSE_TEXT12);
+				if (whatsNextTextWidget != null)
+				{
+					String whatsNextText = whatsNextTextWidget.getText();
+					log.debug("What's next widget text loaded: {}", whatsNextText);
+					if (whatsNextText != null && !whatsNextText.isEmpty())
+					{
+						List<RandomEventItem> initialSelectionItems = this.getPatternNextInitialSelectionMap().values().asList();
+						List<RandomEventItem> choicesItems = this.getPatternNextChoicesMap().values().asList();
+						RandomEventItem answerItem = this.relationshipSystem.findMissingItem(initialSelectionItems, choicesItems);
+						if (answerItem != null)
+						{
+							this.patternNextAnswer = answerItem;
+							Integer interfaceID = getKeyForValue(this.getPatternNextChoicesMap(), answerItem);
+							this.patternNextAnswerWidget = interfaceID != null ? this.client.getWidget(interfaceID) : null;
+							log.debug("Pattern next answer set to: {}", this.patternNextAnswer);
+							log.debug("Pattern next answer widget set to: {}", this.patternNextAnswerWidget);
+						}
+						else
+						{
+							log.warn("No valid answer found for what's next text '{}'.", whatsNextText);
+							this.patternNextAnswer = null;
+							this.patternNextAnswerWidget = null;
+						}
+					}
+					else
+					{
+						log.warn("Next hint widget text is empty or null.");
+						this.patternNextAnswer = null;
+						this.patternNextAnswerWidget = null;
+					}
+				}
+			});
+		}
+	}
+
+	@Subscribe
+	public void onWidgetClosed(WidgetClosed widgetClosed)
+	{
+		// log.debug("Widget closed with group ID: {}", widgetClosed.getGroupId());
+		if (widgetClosed.getGroupId() == InterfaceID.PATTERN_CARDS)
+		{
+			log.debug("Pattern cards widget closed, resetting pattern card answers.");
+			this.patternCardAnswers = null;
+			this.patternCardAnswerWidgets = null;
+		}
+
+		if (widgetClosed.getGroupId() == InterfaceID.PATTERN_NEXT)
+		{
+			log.debug("Pattern next widget closed, resetting pattern next answer.");
+			this.patternNextAnswer = null;
+			this.patternNextAnswerWidget = null;
+		}
+	}
+
+	private ImmutableMultimap<Integer, RandomEventItem> getPatternCardMap()
+	{
+		ImmutableMultimap.Builder<Integer, RandomEventItem> builder = ImmutableMultimap.builder();
+		for (int patternCardSelectionInterfaceID : PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_MODELS)
+		{
+			int modelID = Objects.requireNonNull(this.client.getWidget(patternCardSelectionInterfaceID)).getModelId();
+			RandomEventItem randomEventItem = RandomEventItem.fromModelID(modelID);
+			if (randomEventItem == null)
+			{
+				log.warn("No RandomEventItem found for model ID: {}", modelID);
+				continue;
+			}
+			builder.put(patternCardSelectionInterfaceID, randomEventItem);
+		}
+		return builder.build();
+	}
+
+	private ImmutableMap<Integer, RandomEventItem> getPatternNextInitialSelectionMap()
+	{
+		ImmutableMap.Builder<Integer, RandomEventItem> builder = ImmutableMap.builder();
+		for (int patternNextInitialPatternInterfaceID : PATTERNNEXT_INTERFACEIDS_INITIAL_PATTERN)
+		{
+			int modelID = Objects.requireNonNull(this.client.getWidget(patternNextInitialPatternInterfaceID)).getModelId();
+			RandomEventItem randomEventItem = RandomEventItem.fromModelID(modelID);
+			if (randomEventItem == null)
+			{
+				log.warn("No RandomEventItem found for model ID: {}", modelID);
+				continue;
+			}
+			builder.put(patternNextInitialPatternInterfaceID, randomEventItem);
+		}
+		return builder.build();
+	}
+
+	private ImmutableMultimap<Integer, RandomEventItem> getPatternNextChoicesMap()
+	{
+		ImmutableMultimap.Builder<Integer, RandomEventItem> builder = ImmutableMultimap.builder();
+		for (int patternNextChoiceInterfaceID : PATTERNNEXT_INTERFACEIDS_CHOICES)
+		{
+			int modelID = Objects.requireNonNull(this.client.getWidget(patternNextChoiceInterfaceID)).getModelId();
+			RandomEventItem randomEventItem = RandomEventItem.fromModelID(modelID);
+			if (randomEventItem == null)
+			{
+				log.warn("No RandomEventItem found for model ID: {}", modelID);
+				continue;
+			}
+			builder.put(patternNextChoiceInterfaceID, randomEventItem);
+		}
+		return builder.build();
+	}
+
+	private Integer getKeyForValue(Multimap<Integer, RandomEventItem> map, RandomEventItem value)
+	{
+		for (Map.Entry<Integer, RandomEventItem> entry : map.entries())
+		{
+			if (entry.getValue().equals(value))
+			{
+				return entry.getKey();
+			}
+		}
+		return null;
+	}
+
+	public Widget getPatternCardSelectionWidgetFromModel(Widget modelWidget)
+	{
+		// This method retrieves the widget for the pattern card selection based on the model widget.
+		// PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_MODELS[i] = PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_SELECTS[i]
+		// So if modelWidget corresponds to PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_MODELS[i] then it should return PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_SELECTS[i]
+		for (int i = 0; i < PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_MODELS.length; i++)
+		{
+			if (modelWidget.getId() == PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_MODELS[i])
+			{
+				return this.client.getWidget(PATTERNCARDS_INTERFACEIDS_AVAILABLE_CARD_SELECTS[i]);
+			}
+		}
+		log.warn("No matching selection widget found for model widget ID: {}", modelWidget.getId());
+		return null;
+	}
+}
